shows order and gives total value
